pyproject.toml
[project]
name = "pdb-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.3",
    "biopython",
    "requests",
]
